binder: change nice value of rt to non-rt call
Currently, when a node does not support real-time priority inheritance
but serves a real-time thread, the thread priority is hard coded to 120
with normal scheduling policy.

This patch aims to boost the normal priority of a thread that serves
real-time thread from 120 to 100 to better mitigate a binder priority
inversion on real-time threads.

Bug: 327270040
Test: abtd runs
Change-Id: I8c36edd48a1bc2ede457862cf80533c293c4b0e9
Signed-off-by: tyiu <tyiu@google.com>