abox: Properly advertise abox disable state
This is to prevent the power domain (pd) controlling the abox + display
to try to disable itself when abox is still enabled, which would lead
to issues such as crashes, freezes, or the system refusing to suspend.

Also, in the case of PD trying to disable before Abox was suspended, we
try one more time ~10 ms later to give an extra chance. It can happen
rarely that PD happens at the same time Abox gets disabled, so a retry
mechanism is preferred in order to give an extra chance to disable the
power domain.

Bug: 343048383
Test: Manual
Change-Id: I7c9de090b9b7f336d91c02894b8cc753de9416c8
Signed-off-by: Florian Muller <mullerf@google.com>